Rename helper file now with only multiget code
Single key get() now solely uses kv_helper.h methods
