+++ java/rocksjni/jni_multiget_helpers.h
